// ButorRevolutionWebAPI/src/frontend/web/projects/webshop/src/app/account-history/models/account.ts
export interface IUserBasicInfo {
    name: string;
    phone?: string;
    email: string;
    countryId: number;
    postCode: number;
    city: string;
    address: string;
    languageId: number;
    fileName?: string;
    containerName?: string;
    currentPassword?: string;
    newPassword?: string;
}